Move Notification class to own file
