+++ lib/pushpad/notification.rb
@@ -1,0 +1,71 @@
+module Pushpad
+  class Notification
+    class DeliveryError < RuntimeError
+    end
+
+    attr_accessor :body, :title, :target_url, :icon_url, :ttl, :require_interaction
+
+    def initialize(options)
+      self.body = options[:body]
+      self.title = options[:title]
+      self.target_url = options[:target_url]
+      self.icon_url = options[:icon_url]
+      self.ttl = options[:ttl]
+      self.require_interaction = options[:require_interaction]
+    end
+
+    def broadcast(options = {})
+      deliver req_body(nil, options[:tags]), options
+    end
+
+    def deliver_to(users, options = {})
+      uids = if users.respond_to?(:ids)
+        users.ids
+      elsif users.respond_to?(:collect)
+        users.collect {|u| u.respond_to?(:id) ? u.id : u }
+      else
+        [users.respond_to?(:id) ? users.id : users]
+      end
+      deliver req_body(uids, options[:tags]), options
+    end
+
+    private
+
+    def deliver(req_body, options = {})
+      project_id = options[:project_id] || Pushpad.project_id
+      raise "You must set project_id" unless project_id
+      endpoint = "https://pushpad.xyz/projects/#{project_id}/notifications"
+      uri = URI.parse(endpoint)
+      https = Net::HTTP.new(uri.host, uri.port)
+      https.use_ssl = true
+      req = Net::HTTP::Post.new(uri.path, req_headers)
+      req.body = req_body
+      res = https.request(req)
+      raise DeliveryError, "Response #{res.code} #{res.message}: #{res.body}" unless res.code == '201'
+      JSON.parse(res.body)
+    end
+
+    def req_headers
+      raise "You must set Pushpad.auth_token" unless Pushpad.auth_token
+      {
+        'Authorization' => 'Token token="' + Pushpad.auth_token + '"',
+        'Content-Type' => 'application/json;charset=UTF-8',
+        'Accept' => 'application/json'
+      }
+    end
+
+    def req_body(uids = nil, tags = nil)
+      notification_params = { "body" => self.body }
+      notification_params["title"] = self.title if self.title
+      notification_params["target_url"] = self.target_url if self.target_url
+      notification_params["icon_url"] = self.icon_url if self.icon_url
+      notification_params["ttl"] = self.ttl if self.ttl
+      notification_params["require_interaction"] = self.require_interaction unless self.require_interaction.nil?
+
+      body = { "notification" => notification_params }
+      body["uids"] = uids if uids
+      body["tags"] = tags if tags
+      body.to_json
+    end
+  end
+end
